Use seconds in time array statistics

+++ DevilDaggersInfo.Web.BlazorWasm.Server/Caches/LeaderboardStatistics/ArrayStatistic.cs
@@ -3,10 +3,10 @@ namespace DevilDaggersInfo.Web.BlazorWasm.Server.Caches.LeaderboardStatistics;
 public class ArrayStatistic
 {
 	public double Average { get; private set; }
-	public int Median { get; private set; }
-	public int Mode { get; private set; }
+	public double Median { get; private set; }
+	public double Mode { get; private set; }
 
-	public void Populate(IEnumerable<int> data)
+	public void Populate(IEnumerable<double> data)
 	{
 		data = data.OrderBy(n => n);
